refactor: 아티클 링크 생성기를 분리한다
